Fix tooltip and legend and refactor config
