+++ frontend/docs/mock-data/dashboard-widget-data.json
@@ -1,62 +1,77 @@
 [
   {
     "x": 2003,
-    "y1": 51919439.9942143
+    "y1": 51919439.9942143,
+    "y2": 41919439.9942143
   },
   {
     "x": 2004,
-    "y1": 49549940.9877958
+    "y1": 49549940.9877958,
+    "y2": 39549940.9877958
   },
   {
     "x": 2005,
-    "y1": 51182073.9850648
+    "y1": 51182073.9850648,
+    "y2": 39549940.9877958
   },
   {
     "x": 2006,
-    "y1": 52464639.994372
+    "y1": 52464639.994372,
+    "y2": 32469940.9877958
   },
   {
     "x": 2007,
-    "y1": 57857171.9918152
+    "y1": 57857171.9918152,
+    "y2": 42469940.9877958
   },
   {
     "x": 2008,
-    "y1": 59833104.9916063
+    "y1": 59833104.9916063,
+    "y2": 42469940.9877958
   },
   {
     "x": 2009,
-    "y1": 57345381.9894316
+    "y1": 57345381.9894316,
+    "y2": 42469940.9877958
   },
   {
     "x": 2010,
-    "y1": 70602864.6839387
+    "y1": 70602864.6839387,
+    "y2": 82469940.9877958
   },
   {
     "x": 2011,
-    "y1": 76395808.2945657
+    "y1": 76395808.2945657,
+    "y2": 82469940.9877958
   },
   {
     "x": 2012,
-    "y1": 67629673.6211716
+    "y1": 67629673.6211716,
+    "y2": 37629940.9877958
   },
   {
     "x": 2013,
-    "y1": 83873295.1901526
+    "y1": 83873295.1901526,
+    "y2": 33629940.9877958
   },
   {
     "x": 2014,
-    "y1": 88941394.2096403
+    "y1": 88941394.2096403,
+    "y2": 67629940.9877958
   },
   {
     "x": 2015,
-    "y1": 318190211.140903
+    "y1": 318190211.140903,
+    "y2": 37629940.9877958
   },
   {
     "x": 2016,
-    "y1": 362740274.629392
+    "y1": 362740274.629392,
+    "y2": 37629940.9877958
   },
   {
     "x": 2017,
-    "y1": 402932195.307914
+    "y1": 402932195.307914,
+    "y2": 39629940.9877958
   }
 ]
