+++ frontend/styles/_docz.scss
@@ -64,4 +64,5 @@
 
 .chart-container {
   background-color: $charcoal-dark;
+  padding: 30px;
 }
